Use new environment files in supported runners
The old set-env syntax will be deprecated because of CVE-2020-15228.
A new method of setting environment variables using a file was introduced
to solve this issue.

The SecretHub Action now checks if the runner supports this new method
and uses if it is available. A random heredoc identifier is used to make
sure that it is practically impossible that there is a collision between
the identifier and the content of the secret (even if the content is
controlled by an untrusted source).